removed old api. create api for socket
