+++ package.json
@@ -1,13 +1,14 @@
 {
   "name": "ddk",
   "version": "0.9.9a",
+  "main": "index.js",
   "private": true,
   "scripts": {
     "tsc": "tsc",
-    "start": "npm run start:core",
-    "build": "npm run build:core",
+    "build": "npm run build:core & npm run build:api",
     "start:api": "node --max-old-space-size=8192 dist/api/app.js",
-    "start:core": "npm run tsc && node --max-old-space-size=8192 dist/core/app.js",
+    "start:core": "node --max-old-space-size=8192 dist/core/app.js",
+    "server": "npm run server:core & npm run server:api",
     "server:api": "nodemon --max-old-space-size=8192 dist/api/app.js",
     "server:core": "nodemon --max-old-space-size=8192 dist/core/app.js",
     "debug": "node --inspect=0.0.0.0:9229 app.js",
@@ -15,11 +16,12 @@
     "server-docs": "npm run jsdoc && http-server docs/jsdoc/",
     "build:api": "./node_modules/webpack/bin/webpack.js --env.service=api --progress --profile --colors --display-modules",
     "build:core": "./node_modules/webpack/bin/webpack.js --env.service=core --progress --profile --colors --display-modules",
+    "watch": "npm run watch:core & npm run watch:api",
     "watch:api": "./node_modules/webpack/bin/webpack.js --env.service=api --progress --colors --watch",
     "watch:core": "./node_modules/webpack/bin/webpack.js --env.service=core --progress --colors --watch",
     "clean": "rm -rf node_modules/ dist/ dapps/ .nyc_output/ logs/ .npm/ .config/ .node-gyp",
-    "watch:with_update_clean": "npm run clean && npm install && npm run watch",
     "watch:with_update": "npm install && npm run watch",
+    "watch:with_update_clean": "npm run clean && npm install && npm run watch",
     "test": "export NODE_PATH=./ && mocha --require ts-node/register './test/**/*.test.ts'",
     "test:decorator": "export NODE_PATH=./ && mocha --require ts-node/register './test/shared/decorator/*.test.ts'",
     "test:migrate": "export NODE_PATH=./ && mocha --require ts-node/register './test/shared/migrate/*.test.ts'",
