+++ tsconfig.json
@@ -12,7 +12,7 @@
     "outDir": "dist",
     "baseUrl": ".",
     "paths": {
-//      "api/*": ["./api/*"],
+      "api/*": ["./api/*"],
       "core/*": ["./core/*"],
       "shared/*": ["./shared/*"],
       "config/*": ["./config/*"]
@@ -27,7 +27,7 @@
     ]
   },
   "include": [
-//    "api/**/*",
+    "api/**/*",
     "core/**/*",
     "shared/**/*",
     "test/**/*",
